Make codility compatible.
